Fix invalid ports in the Debian Stretch template
Debian stretch does not have an imap3 entry in /etc/service, but this
port is referenced in the default jail.conf configuration file.

THe problem was addressed upstream, but the fix was not rolled-out to
the Debian package:
https://github.com/fail2ban/fail2ban/issues/1942

Fix this locally so that these jails do not fail to start.